+++ pkg/telemetry/telemetry_test.go
@@ -90,6 +90,7 @@ func TestHeartbeat(t *testing.T) {
 			tr := &TelemetryClient{
 				RWMutex:    sync.RWMutex{},
 				properties: tt.fields.properties,
+				profile:    NewNoopPerfProfile(),
 			}
 			tr.heartbeat(tt.args.ctx)
 		})
@@ -125,6 +126,7 @@ func TestTelemetryClient_StopPerf(t *testing.T) {
 			tr := &TelemetryClient{
 				RWMutex:    sync.RWMutex{},
 				properties: tt.fields.properties,
+				profile:    NewNoopPerfProfile(),
 			}
 			tr.StopPerf(tt.args.counter)
 		})
